Add editqf_no_type_mappings to disable "types"
W and E are common motion commands that may be used in the quickfix
window, so add an option to prevent them from being mapped. The second
list of mappings we always want to map because they don't make sense in
the quickfix window (which is not modifiable).

For consistency, we also exclude I.